- Compile on Ubuntu 14.04 - Clean up src/type.h :   * Avoid type `long', its size varies among host architectures   * The types uint / ushort seem to be predefined on Ubuntu     (may need tweaking on other platforms) - Update file headers (copyright / contact) - In src/version.h set SIM1750_VERSION to "2.3c"

+++ sim1750/src/break.h
@@ -7,7 +7,7 @@
 
 typedef enum { READ_WRITE, READ, WRITE } breaktype;
 
-extern int  find_breakpt (breaktype type, ulong phys_address);
+extern int  find_breakpt (breaktype type, uint phys_address);
 extern void set_inactive (int bp_index);
 extern void set_active   (int bp_index);
