+++ sim1750/src/smemacc.h
@@ -2,7 +2,7 @@
 
 #include "type.h"
 
-extern ulong get_phys_address (int bank, ushort as, ushort log_addr);
+extern uint get_phys_address (int bank, ushort as, ushort log_addr);
 extern bool  get_raw (int bank, ushort as, ushort address, ushort *value);
 extern void  store_raw (int bank, ushort as, ushort address, ushort value);
